% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contour_sep_distance.R
\name{get_contour_sep_dist}
\alias{get_contour_sep_dist}
\title{get_contour_sep_dist}
\usage{
get_contour_sep_dist(
  file_grd,
  epsg = 32632,
  levels = NULL,
  x_source = 738259,
  y_source = 5073666,
  degree_step = 5,
  trans_factor = 1
)
}
\arguments{
\item{file_grd}{string path to the CALPOST grid file}

\item{epsg}{number epsg to set crs in the raster object}

\item{levels}{vector of the levels for the contourplot}

\item{x_source}{x coordinates of the source (origin) point}

\item{y_source}{y coordinates of the source (origin) point}

\item{degree_step}{amplitude of angle degrees to scan direction}

\item{trans_factor}{number, transformation factor to be applied to calculated values, leave it as default = 1}
}
\value{
df
}
\description{
Get sepration distances from contourlines
}
