% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpost_grd.R
\name{mapview_calpost_grd}
\alias{mapview_calpost_grd}
\title{mapview_calpost_grd}
\usage{
mapview_calpost_grd(
  file_grd,
  epsg = 32632,
  at = NULL,
  export = FALSE,
  name_of_map_layer = "mapview_layer_name",
  string_filename = "export_filename",
  trans_factor = 1
)
}
\arguments{
\item{file_grd}{string path to the CALPOST grid file}

\item{epsg}{number epsg to set crs in the raster object}

\item{at}{numeric vector defining breakpoints for the visualization,
default value is NULL by using continuous 'pretty' breakpoints, see lattice::levelplot function for details}

\item{export}{logical, export mapview object as hmtl and png files? The default value is equal to FALSE}

\item{name_of_map_layer}{string to name the mapview layer}

\item{string_filename}{string to name the output files}

\item{trans_factor}{number, transformation factor to be applied to calculated values, leave it as default = 1}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from CALPOST grid file and export as html and png files
}
