% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpuff_setup_qafiles.R
\name{mapview_calpuff_terr}
\alias{mapview_calpuff_terr}
\title{mapview_qaterr_grd}
\usage{
mapview_calpuff_terr(
  file_grd,
  epsg = 32632,
  string_filename = " file_name_grid_terr",
  name_of_map_layer = "layer_name_grid_terr"
)
}
\arguments{
\item{file_grd}{string path to the Calpuff setup grid file qaterr.grd}

\item{epsg}{number epsg to set crs in the raster object}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from Calpuff setup grid file qaterr.grd and export as html and png files
}
