% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpuff_setup_qafiles.R
\name{mapview_emi_points}
\alias{mapview_emi_points}
\title{mapview_emi_points}
\usage{
mapview_emi_points(
  qapnts.dat,
  epsg = 32632,
  string_filename = "file_name_emi_points",
  name_of_map_layer = "layer_name_emi_points"
)
}
\arguments{
\item{qapnts.dat}{string path to Calpuff setup file emission points qapnts.dat}

\item{epsg}{number epsg to set crs of the emission points}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
mapview Calpuff setup emission points
}
