% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_grd.R
\name{mapview_grd}
\alias{mapview_grd}
\title{mapview_grd}
\usage{
mapview_grd(file_grd, epsg, string_filename, name_of_map_layer)
}
\arguments{
\item{file_grd}{string path to the Calpuff input file name *.grd}

\item{epsg}{epsg EPSG number for setting raster crs}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{a string to name the layer of the map}
}
\description{
Create mapview object from Calpuff grd file, export as html and png
}
