% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_setup_qafiles.R
\name{mapview_met_domain}
\alias{mapview_met_domain}
\title{mapview_met_domain}
\usage{
mapview_met_domain(qagrid_bna, epsg, string_filename, name_of_map_layer)
}
\arguments{
\item{qagrid_bna}{string path to Calpuff setup file qagrid.bna}

\item{epsg}{number epsg to set crs of the domain}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
Read Calpuff setup meteorological domain
}
