% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpuff_setup_qafiles.R
\name{mapview_samp_domain}
\alias{mapview_samp_domain}
\title{mapview_samp_domain}
\usage{
mapview_samp_domain(
  qagrid.bna,
  epsg = 32632,
  string_filename = "file_name_samp_domain",
  name_of_map_layer = "layer_name_samp_domain"
)
}
\arguments{
\item{qagrid.bna}{string path to Calpuff setup file qagrid.bna}

\item{epsg}{number epsg to set crs of the file domain}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
mapview Calpuff setup sampling domain
}
