% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpuff_setup_qafiles.R
\name{mapview_terr_grd}
\alias{mapview_terr_grd}
\title{Mapview CALPUFF terrain}
\usage{
mapview_terr_grd(
  qaterr.grd,
  epsg = 32632,
  export = FALSE,
  name_of_map_layer = "layer_name_grid_terr",
  string_filename = " file_name_grid_terr"
)
}
\arguments{
\item{qaterr.grd}{string path to the Calpuff setup grid file 'qaterr.grd'}

\item{epsg}{number epsg to set crs in the raster object}

\item{export}{logical, export mapview object as hmtl and png files? The default value is equal to FALSE}

\item{name_of_map_layer}{string to name the mapview layer}

\item{string_filename}{string to name the output files}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from CALPUFF setup grid file 'qaterr.grd' and eventually export it as html and png files
}
