% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_setup_qafiles.R
\name{read_qagrid_bna}
\alias{read_qagrid_bna}
\title{read_qagrid_bna}
\usage{
read_qagrid_bna(qagrid.bna, epsg = 32632, skip, n_max)
}
\arguments{
\item{qagrid.bna}{string path to Calpuff setup file qagrid.bna}

\item{epsg}{number epsg to set crs of the file domain}

\item{skip}{number of rows to skip}

\item{n_max}{max number of rows to read}
}
\value{
sf object Calpuff domain
}
\description{
Read Calpuff setup domain
}
\keyword{internal}
