% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_setup_qafiles.R
\name{read_qapnts_dat}
\alias{read_qapnts_dat}
\title{read_qapnts_dat}
\usage{
read_qapnts_dat(qapnts.dat, epsg = 32632)
}
\arguments{
\item{qapnts.dat}{string path to Calpuff setup file emission points qapnts.dat}

\item{epsg}{number epsg to set crs of the emission points}
}
\value{
sf object Calpuff emission points
}
\description{
Read Calpuff setup emission points
}
\keyword{internal}
