% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_setup_qafiles.R
\name{read_qarecd_dat}
\alias{read_qarecd_dat}
\title{read_qarecd_dat}
\usage{
read_qarecd_dat(qarecd.dat, epsg = 32632)
}
\arguments{
\item{qarecd.dat}{string path to Calpuff setup file discrete receptors qarecd.dat}

\item{epsg}{number epsg to set crs of the discrete receptors}
}
\value{
sf object Calpuff discrete receptors
}
\description{
Read Calpuff setup discrete receptors
}
\keyword{internal}
