% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_setup_qafiles.R
\name{read_qarecg_dat}
\alias{read_qarecg_dat}
\title{read_qarecg_dat}
\usage{
read_qarecg_dat(qarecg.dat, epsg = 32632)
}
\arguments{
\item{qarecg.dat}{string path to Calpuff setup file grid receptors qarecg.dat}

\item{epsg}{number epsg to set crs of the grid receptors}
}
\value{
sf object Calpuff grid receptors
}
\description{
Read Calpuff setup grid receptors
}
\keyword{internal}
