% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_setup_qafiles.R
\name{read_qaterr_grd}
\alias{read_qaterr_grd}
\title{read_qaterr_grd}
\usage{
read_qaterr_grd(qaterr.grd, epsg = 32632)
}
\arguments{
\item{qaterr.grd}{string path to Calpuff setup terrain grid file}

\item{epsg}{number epsg to set crs in the raster object}
}
\value{
raster object
}
\description{
Read Calpuff setup file terrain grid
}
\keyword{internal}
