% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_calpuff_setup_qafiles.R
\name{shapefile_disc_recs}
\alias{shapefile_disc_recs}
\title{shapefile_disc_recs}
\usage{
shapefile_disc_recs(
  qarecd.dat,
  epsg = 32632,
  string_filename = "file_name_disc_recs"
)
}
\arguments{
\item{qarecd.dat}{string path to Calpuff setup file discrete receptors qarecd.dat}

\item{epsg}{number epsg to set crs of the discrete receptors}

\item{string_filename}{string to name the output files}
}
\value{
shapefile of the discrete receptors
}
\description{
shapefile Calpuff setup discrete receptors
}
