% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_calpuff_setup_qafiles.R
\name{shapefile_emi_points}
\alias{shapefile_emi_points}
\title{shapefile_emi_points}
\usage{
shapefile_emi_points(
  qapnts.dat,
  epsg = 32632,
  string_filename = "file_name_emi_points"
)
}
\arguments{
\item{qapnts.dat}{string path to Calpuff setup file emission points qapnts.dat}

\item{epsg}{number epsg to set crs of the emission points}

\item{string_filename}{string to name the output files}
}
\value{
shapefile of the emission points
}
\description{
shapefile Calpuff setup emission points
}
