% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_calpuff_setup_qafiles.R
\name{shapefile_grid_recs}
\alias{shapefile_grid_recs}
\title{shapefile_grid_recs}
\usage{
shapefile_grid_recs(
  qarecg.dat,
  epsg = 32632,
  string_filename = "file_name_grid_recs"
)
}
\arguments{
\item{qarecg.dat}{string path to Calpuff setup file grid receptors qarecg.dat}

\item{epsg}{number epsg to set crs of the grid receptors}

\item{string_filename}{string to name the output files}
}
\value{
shapefile of the grid receptors
}
\description{
shapefile Calpuff setup grid receptors
}
