% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contour_sep_distance.R
\name{get_contour_sep_dist}
\alias{get_contour_sep_dist}
\title{Separation distances from contour lines}
\usage{
get_contour_sep_dist(
  file_grd,
  epsg = 32632,
  x_source,
  y_source,
  levels = NULL,
  degree_step = 5,
  trans_factor = 1
)
}
\arguments{
\item{file_grd}{string path to the CALPOST grid file}

\item{epsg}{number epsg to set crs in the raster object}

\item{x_source}{x coordinates of the source (origin) point}

\item{y_source}{y coordinates of the source (origin) point}

\item{levels}{vector of defined levels for the contour lines}

\item{degree_step}{angle degree to scan directions where calculate distances from source (origin) point}

\item{trans_factor}{number, transformation factor to be applied to calculated values, leave it as default = 1}
}
\value{
a simple feature data frame
}
\description{
Get a a simple feature data frame of separation distances from the source
(origin) point to the concentration contour lines
}
