% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpost_grd.R
\name{mapview_calpost_grd}
\alias{mapview_calpost_grd}
\title{mapview_calpost_grd}
\usage{
mapview_calpost_grd(
  file_grd,
  epsg = 32632,
  string_filename = " file_name",
  name_of_map_layer = "layer_name"
)
}
\arguments{
\item{file_grd}{string path to the Calpost grid file}

\item{epsg}{number epsg to set crs in the raster object}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from Calpost grid file and export as html and png files
}
