% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpuff_setup_qafiles.R
\name{mapview_comp_domain}
\alias{mapview_comp_domain}
\title{Mapview CALPUFF computational domain}
\usage{
mapview_comp_domain(
  qagrid.bna,
  epsg = 32632,
  export = FALSE,
  string_filename = "file_name_comp_domain",
  name_of_map_layer = "layer_name_comp_domain"
)
}
\arguments{
\item{qagrid.bna}{string path to Calpuff setup file 'qagrid.bna'}

\item{epsg}{number epsg to set crs of the file domain}

\item{export}{logical, export mapview object as hmtl and png files? The default value is equal to FALSE}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from CALPUFF setup computational domain file 'qagrid.bna' and eventually export it as html and png files
}
