% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpuff_setup_qafiles.R
\name{mapview_disc_recs}
\alias{mapview_disc_recs}
\title{mapview_disc_recs}
\usage{
mapview_disc_recs(
  qarecd.dat,
  epsg = 32632,
  string_filename = "file_name_disc_rescs",
  name_of_map_layer = "layer_name_disc_recs"
)
}
\arguments{
\item{qarecd.dat}{string path to Calpuff setup file discrete receptors qarecd.dat}

\item{epsg}{number epsg to set crs of the discrete receptors}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
mapview Calpuff setup discrete receptors
}
