% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_calpuff_setup_qafiles.R
\name{mapview_emi_points}
\alias{mapview_emi_points}
\title{Mapview CALPUFF emission points}
\usage{
mapview_emi_points(
  qapnts.dat,
  epsg = 32632,
  export = FALSE,
  string_filename = "file_name_emi_points",
  name_of_map_layer = "layer_name_emi_points"
)
}
\arguments{
\item{qapnts.dat}{string path to Calpuff setup file emission points 'qapnts.dat'}

\item{epsg}{number epsg to set crs of the emission points}

\item{export}{logical, export mapview object as hmtl and png files? The default value is equal to FALSE}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from CALPUFF setup emission points file 'qapnts.dat' and eventually export it as html and png files
}
