% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapview_grd.R
\name{mapview_grd}
\alias{mapview_grd}
\title{mapview_grd}
\usage{
mapview_grd(file_grd, epsg, string_filename, name_of_map_layer)
}
\arguments{
\item{file_grd}{string path to the Calpuff input file name *.grd}

\item{epsg}{number epsg to set crs in the raster object}

\item{string_filename}{string to name the output files}

\item{name_of_map_layer}{string to name the mapview layer}
}
\value{
mapview object exported as html and png files
}
\description{
Create a mapview object from Calpuff grd file and export as html and png files
}
