% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_calpost_grd.R
\name{read_calpost_grd}
\alias{read_calpost_grd}
\title{read_calpost_grd}
\usage{
read_calpost_grd(file_grd, epsg = 32632)
}
\arguments{
\item{file_grd}{string path to Calpost grid file}

\item{epsg}{number epsg to set crs in the raster object}
}
\value{
raster object
}
\description{
Read Calpost grid file
}
\keyword{internal}
