% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grd.R
\name{read_grd}
\alias{read_grd}
\title{read_grd}
\usage{
read_grd(file_grd, epsg)
}
\arguments{
\item{file_grd}{Calpuff grid file}

\item{epsg}{EPSG number for setting raster crs}
}
\value{
raster object
}
\description{
Read Calpuff output grid files
}
